+++ src/containers/HierarchyBuilder/FolderHierarchy/FolderHierarchy.styled.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  width: 100%;
+  justify-content: flex-end;
+`
